Optimizations and bugfixes
- Removed some of the dead code
- Removed un-needed string mallocs from bison parser
- Reverted NUM to double
- Fixed a mistake where multiple AND's or OR's without precedence gave invalid syntax
- Sanity checks in str_duplicate and str_concat

Commit by AvyChanna on 22/03/2020 07:29 PM

+++ Lab-3/sql_runner.h
@@ -14,7 +14,7 @@ typedef struct ast {
 	// (those with logical (>= <= > < = <> !=) operators)
 	int operand_type[2];		// int str var
 	char *str[2];				// salary name
-	int num[2];					// 10 20
+	double num[2];					// 10 20
 } ast;
 
 typedef struct clq {
